Create Hello World
write a program that will print the hello world. If you want to print the hello world five times then just write the print statement more four times and it will shows the five hello world in a line. in order to make it in a separate line just using slash n (\n) after the world.